delete mutations pulled out into seperate instances to solve 404 errors from invalidation fetching
